Add attribution to nomics.com
