+++ src/components/Attribution.js
@@ -1,0 +1,8 @@
+import { Link } from '@material-ui/core';
+import React from 'react';
+
+const Attribution = () => (
+  <Link href="https://nomics.com">Crypto Market Cap & Pricing Data Provided By Nomics</Link>
+);
+
+export default Attribution;
